Got bencode to work in Py3K.
A round-trip thorugh the bencoding system as it now stands will
attempt to preserve string keys in dicts as strings, but all other
strings will emerge as bytestrings decoded from UTF8. Also, boolean
values will emerge as ints.

Otherwise this should work as a drop in replacement. As it encodes
to bytes, it may limit the changes needed in other portions of the
code (networking for example..